% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat2Scanpy.R
\name{CreateScanpyObject}
\alias{CreateScanpyObject}
\title{Interchangeable single-cell data formats}
\usage{
CreateScanpyObject(seu, save_name, save_embedding = TRUE)
}
\arguments{
\item{seu}{\code{seu} is the seurat object to be converted}

\item{save_name}{\code{save_name} is the prefix of the output file's name.}

\item{save_embedding}{\code{save_embedding} shoule be set to TRUE when you
hope put embeddings (e.g. PCA, UMAP) into the converted file.}
}
\value{
No value returned. an h5ad file is written to the disk.
}
\description{
Convert a Seurat object to a scanpy object.
}
\details{
This R function accepts a Seurat object and write a scanpy h5ad file to the disk.

multiple parameters could be specified to control the conversion behaviour.
}
\author{
Sijie Chen
}
